style: move tracking fn to new file
