+++ lti_consumer/lti_1p1/contrib/tests/test_django.py
@@ -31,7 +31,7 @@ class TestLtiEmbed(TestCase):
         # Patch calls to LMS event tracking
         self._mock_emit_track_event = Mock()
         track_event_patcher = patch(
-            'lti_consumer.utils.get_event_tracker',
+            'lti_consumer.track.get_event_tracker',
             return_value=Mock(emit=self._mock_emit_track_event),
         )
         self.addCleanup(track_event_patcher.stop)
